respect tty discovery for CLICOLOR specs
This commit also:

* fixes the code to prioritise CLICOLOR_FORCE over NO_COLOR.
The docs said that there is this preference, but in fact
the opposite preference took place before.
(Pattern match had wrong order of lines.)

* removes the prioritization example table because it was wrong,
and it may be harder to read and maintain than the text description above.
(It was wrong on line 5.)

* fix a specs mismatch in handling `CLICOLOR_FORCE`.
A value of `0` should be treated identically to unset.

* make `ShouldColorize.clicolor` be `bool` instead of `Option<bool>`,
because the optional boolean does not hold any meaning really,
we always practically unwrap it as `.unwrap_or_else(|| true)`.

* fix and add tests in accordance to the specs

fixes GH-54